refactor: change app struct and added routes
