+++ lib/src/app.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+
+import 'core/router.dart';
+import 'core/ui/design_system/design_system.dart';
+import 'core/ui/widgets/keyboard_dismiss_on_tap.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp.router(
+      title: 'My School',
+      theme: AppTheme.theme,
+      routerConfig: AppRouter.router,
+      debugShowCheckedModeBanner: false,
+      locale: const Locale('pt', 'BR'),
+      supportedLocales: const [Locale('pt', 'BR')],
+      localizationsDelegates: const [
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+        GlobalCupertinoLocalizations.delegate,
+      ],
+      builder: (context, child) => KeyboardDismissOnTap(
+        child: child!,
+      ),
+    );
+  }
+}
