+++ lib/src/modules/home/ui/pages/home_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+
+  static String get route => '/';
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold();
+  }
+}
